refactor(profile): smaller list on mobile
