+++ src/containers/Profile/index.tsx
@@ -1,46 +1,14 @@
 import { useState } from 'react';
-import classNames from 'classnames';
-import { PermissionApp } from 'types/Enums';
-import { useUser, useHavePermission } from 'api/hooks/User';
-import URLS from 'URLS';
-import { useNavigate } from 'react-router-dom';
+import { useUser } from 'api/hooks/User';
 import Helmet from 'react-helmet';
-import { useLogout } from 'api/hooks/User';
 
 // Material-UI
-import {
-  makeStyles,
-  useMediaQuery,
-  Typography,
-  Button,
-  SvgIconProps,
-  Badge,
-  Collapse,
-  List,
-  ListItem,
-  ListItemIcon,
-  ListItemText,
-  Theme,
-} from '@material-ui/core';
+import { makeStyles, useMediaQuery, Typography, Button, Theme } from '@material-ui/core';
 import Skeleton from '@material-ui/lab/Skeleton';
-
-// Icons
-import EventIcon from '@material-ui/icons/DateRangeRounded';
-import NotificationsIcon from '@material-ui/icons/NotificationsNoneRounded';
-import SettingsIcon from '@material-ui/icons/SettingsRounded';
-import AdminIcon from '@material-ui/icons/TuneRounded';
-import LogOutIcon from '@material-ui/icons/ExitToAppRounded';
-import BadgesIcon from '@material-ui/icons/EmojiEventsRounded';
-import GroupsIcon from '@material-ui/icons/PeopleOutlineRounded';
 
 // Project Components
 import Navigation from 'components/navigation/Navigation';
-import ProfileAdmin from 'containers/Profile/components/ProfileAdmin';
-import ProfileSettings from 'containers/Profile/components/ProfileSettings';
-import ProfileEvents from 'containers/Profile/components/ProfileEvents';
-import ProfileGroups from 'containers/Profile/components/ProfileGroups';
-import ProfileNotifications from 'containers/Profile/components/ProfileNotifications';
-import ProfileBadges from 'containers/Profile/components/ProfileBadges';
+import ProfileContent from 'containers/Profile/components/ProfileContent';
 import Paper from 'components/layout/Paper';
 import Dialog from 'components/layout/Dialog';
 import Avatar from 'components/miscellaneous/Avatar';
@@ -76,8 +44,6 @@ const useStyles = makeStyles((theme) => ({
   },
   skeleton: {
     animation: 'animate 1.5s ease-in-out infinite',
-  },
-  skeletonText: {
     margin: 'auto',
     minHeight: 35,
   },
@@ -85,73 +51,13 @@ const useStyles = makeStyles((theme) => ({
     margin: `${theme.spacing(0.25)}px auto`,
     color: theme.palette.text.primary,
   },
-  memberProof: {
-    display: 'flex',
-    justifyContent: 'center',
-  },
-  content: {
-    display: 'grid',
-    gridTemplateColumns: '250px 1fr',
-    gridGap: theme.spacing(1),
-    marginTop: theme.spacing(-6),
-    marginBottom: theme.spacing(3),
-    [theme.breakpoints.down('sm')]: {
-      gridTemplateColumns: '1fr',
-    },
-  },
-  contentList: {
-    overflow: 'hidden',
-  },
-  list: {
-    padding: theme.spacing(0),
-  },
-  logOutButton: {
-    color: theme.palette.error.main,
-  },
 }));
 
 const Profile = () => {
   const classes = useStyles();
-  const navigate = useNavigate();
-  const logOut = useLogout();
   const { data: user } = useUser();
-  const { allowAccess: isAdmin } = useHavePermission([PermissionApp.EVENT, PermissionApp.JOBPOST, PermissionApp.NEWS, PermissionApp.USER]);
   const [showModal, setShowModal] = useState(false);
   const xsDown = useMediaQuery((theme: Theme) => theme.breakpoints.down('sm'));
-
-  const logout = () => {
-    logOut();
-    navigate(URLS.landing);
-  };
-
-  const eventTab = { label: 'Arrangementer', icon: EventIcon };
-  const notificationsTab = { label: 'Varsler', icon: NotificationsIcon, badge: user?.unread_notifications };
-  const badgesTab = { label: 'Badges', icon: BadgesIcon };
-  const groupsTab = { label: 'Grupper', icon: GroupsIcon };
-  const settingsTab = { label: 'Innstillinger', icon: SettingsIcon };
-  const adminTab = { label: 'Admin', icon: AdminIcon };
-  const logoutTab = { label: 'Logg ut', icon: LogOutIcon, onClick: logout, className: classes.logOutButton };
-  const tabs = [eventTab, notificationsTab, badgesTab, groupsTab, settingsTab, ...(isAdmin ? [adminTab] : []), logoutTab];
-  const [tab, setTab] = useState(eventTab.label);
-
-  type NavListItem = {
-    label: string;
-    icon: React.ComponentType<SvgIconProps>;
-    badge?: string | number;
-    onClick?: () => void;
-    className?: string;
-  };
-
-  const NavListItem = ({ label, icon: Icon, onClick, badge, className = '', ...props }: NavListItem) => (
-    <ListItem button onClick={onClick ? onClick : () => setTab(label)} selected={tab === label} {...props}>
-      <ListItemIcon>
-        <Badge badgeContent={badge} color='error'>
-          <Icon className={className} color={tab === label ? 'primary' : 'inherit'} />
-        </Badge>
-      </ListItemIcon>
-      <ListItemText primary={label} />
-    </ListItem>
-  );
 
   return (
     <Navigation banner={<div className={classes.top} />} fancyNavbar>
@@ -180,46 +86,16 @@ const Profile = () => {
             </>
           ) : (
             <>
-              <Skeleton className={classNames(classes.skeleton, classes.skeletonText)} variant='text' width='75%' />
-              <Skeleton className={classNames(classes.skeleton, classes.skeletonText)} variant='text' width='45%' />
-              <Skeleton className={classNames(classes.skeleton, classes.skeletonText)} variant='text' width='35%' />
+              <Skeleton className={classes.skeleton} variant='text' width='75%' />
+              <Skeleton className={classes.skeleton} variant='text' width='45%' />
+              <Skeleton className={classes.skeleton} variant='text' width='35%' />
             </>
           )}
           <Button className={classes.button} color='primary' onClick={() => setShowModal(true)} variant='contained'>
             Medlemsbevis
           </Button>
         </Paper>
-        <div className={classes.content}>
-          <div>
-            <Paper className={classes.contentList} noPadding>
-              <List aria-label='Profil innholdsliste' className={classes.list}>
-                {tabs.map((tab) => (
-                  <NavListItem {...tab} key={tab.label} />
-                ))}
-              </List>
-            </Paper>
-          </div>
-          <div>
-            <Collapse in={tab === eventTab.label}>
-              <ProfileEvents />
-            </Collapse>
-            <Collapse in={tab === notificationsTab.label} mountOnEnter unmountOnExit>
-              <ProfileNotifications />
-            </Collapse>
-            <Collapse in={tab === badgesTab.label} mountOnEnter>
-              <ProfileBadges />
-            </Collapse>
-            <Collapse in={tab === groupsTab.label} mountOnEnter>
-              <ProfileGroups />
-            </Collapse>
-            <Collapse in={tab === settingsTab.label} mountOnEnter>
-              <Paper>{user && <ProfileSettings user={user} />}</Paper>
-            </Collapse>
-            <Collapse in={tab === adminTab.label} mountOnEnter>
-              <ProfileAdmin />
-            </Collapse>
-          </div>
-        </div>
+        <ProfileContent />
       </div>
     </Navigation>
   );
